Windows: Override enacl compilation on Windows
As is enacl can be compiled on bsd, linux, mac and solaris.

When trying to compile under Windows it will just skip the compile phase.

A separate PR for enacl has been sent: https://github.com/jlouis/enacl/pull/23,
but until it will be reviewed and applied, this change will override
the targets under windows and allow the targets to be compiled

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>