disables erldns cache to improve latency in resolving the DNS query
It was observed that when a new task is launched then it would
take close to 5 seconds to resolve the DNS even though erldns
had been updated with the new record. This was happening because
of the erldns cache. This patch fixes the issue by disabling the
erldns cache. This might have an impact on erldns but we believe
it should be minimal as all the queries are local to an agent.