Allow agent hostnames to be real DNS names
The hostname value of a host can be a real DNS name if you use
`--hostname` or `--hostname-lookup` when starting master and agents.

This was assuming that the hostname was an IP address and you would get
an unhandled einval exception when calling the inet ipv4 parse.

This changes this to properly resolve a hostname and handle the new
error cases, such as nxdomain and einval