Fix namespace and netns types.
There are many functions with NetNs arguments, but some of them expect
`netns()` while others expect a key of `netns(). It's a mess, but it
would require huge refactoring to make the naming better. This commit is
fixing only `dcos_l4lb_mgr` module.

Also, there was a minor issue when networking namespaces are removed.